+++ apps/explorer/test/explorer/chain/import/runner/internal_transactions_test.exs
@@ -288,6 +288,29 @@ defmodule Explorer.Chain.Import.Runner.InternalTransactionsTest do
       assert %{consensus: true} = Repo.get(Block, full_block.hash)
       assert PendingBlockOperation |> Repo.get(full_block.hash) |> is_nil()
     end
+
+    test "does not overwrite gas_used from transaction" do
+      block = insert(:block)
+
+      transaction =
+        insert(:transaction,
+          gas_used: 30_000,
+          cumulative_gas_used: 30_000,
+          index: 0,
+          block_hash: block.hash,
+          block_number: block.number
+        )
+
+      insert(:pending_block_operation, block_hash: block.hash, fetch_internal_transactions: true)
+
+      transaction_changes = make_internal_transaction_changes(transaction, 0, nil)
+
+      assert {:ok, _} = run_internal_transactions([transaction_changes])
+
+      assert Repo.get(Transaction, transaction.hash).gas_used == Decimal.new(30_000)
+      assert %{consensus: true} = Repo.get(Block, block.hash)
+      assert is_nil(Repo.get(PendingBlockOperation, block.hash))
+    end
   end
 
   defp run_internal_transactions(changes_list, multi \\ Multi.new()) when is_list(changes_list) do